labs/lab05/pyproject.toml
[project]
name = "lab05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.3"
dependencies = [
    "apache-airflow==2.8.1",
    "apache-airflow-providers-postgres>=5.14.0",
    "faker>=40.5.1",
    "pyarrow>=23.0.1",
]
